Merge Development to Master ok ready to master

+++ app/Http/Livewire/Template/AddAction.php
@@ -92,6 +92,7 @@ class AddAction extends Component
      */
     public function update()
     {
+        $this->authorize('UPDATE_CONTENT_USR', $this->template->user_id);
         $this->validate();
 
         if ($this->link_attachment == '') {
@@ -128,7 +129,7 @@ class AddAction extends Component
      */
     public function delete()
     {
-
+        $this->authorize('DELETE_CONTENT_USR', $this->template->user_id);
         Action::destroy($this->actionId);
         $this->actionId = null;
         $this->confirmingActionRemoval = false;
