+++ app/Http/Livewire/Template/OneWay.php
@@ -37,6 +37,7 @@ class OneWay extends Component
         $this->validate();
 
         $template = Template::findOrFail($this->selectedItemId);
+        $this->authorize('UPDATE_CONTENT_USR', $template->user_id);
         $template->update($this->modelData());
 
         $this->modalActionVisible = false;
@@ -47,6 +48,7 @@ class OneWay extends Component
     public function edit($id)
     {
         $template = Template::findOrFail($id);
+        $this->authorize('UPDATE_CONTENT_USR', $template->user_id);
         $this->selectedItemId = $id;
         $this->type = $template->type;
         $this->name = $template->name;
@@ -59,6 +61,7 @@ class OneWay extends Component
     public function delete($id)
     {
         $template = Template::findOrFail($id);
+        $this->authorize('DELETE_CONTENT_USR', $template->user_id);
         $template->delete();
 
         $this->emit('refreshLivewireDatatable');
